Add files via upload
The name of the app is Jecky Bakery. It was developed using Ionic, Angular, TypeScript, HTML and CSS. It uses arrays as its storage. The first array is the main array which contains all available product. Then you can select a product and add it to a Cart which is a different array. In the Cart you have two options to either delete the Item or go to CheckOut to order the product, enter the delivery method, payment method and confirm your pricing. There's also an option for the Profile where by there's a default picture, should the user decide to change the picture they can access the device' camera by clicking on the picture and take new picture.There is a log out button in the profile page which will return the user to the login page which is the landing page for the app. The Login page and Registration page are soon to be integrated with verification functions once the database is up and running.